+++ resources/views/livewire/tour-overlay.blade.php
@@ -1,12 +1,14 @@
 <div class="io-overlay" data-tour-overlay>
     @if ($tour && $payload)
         @if ($tour->isChangelog())
-            <x-infinito-onboarding::changelog-modal
-                :id="'io-changelog-' . $tour->id"
-                :tours="[$tour]"
-                :heading="$tour->name"
-                :auto-open="true"
-            />
+            <div x-data x-init="$wire.track('view')">
+                <x-infinito-onboarding::changelog-modal
+                    :id="'io-changelog-' . $tour->id"
+                    :tours="[$tour]"
+                    :heading="$tour->name"
+                    :auto-open="true"
+                />
+            </div>
         @else
             <div
                 wire:key="io-tour-{{ $tour->id }}-{{ $tour->version }}"
@@ -18,6 +20,7 @@
                     labels: @js($labels),
                     onCompleted: () => $wire.markCompleted(),
                     onDismissed: () => $wire.markDismissed(),
+                    onEvent: (name, detail) => $wire.track(name, detail),
                 })"
             ></div>
         @endif